// master-thesis-app/generated/client/enums.ts
export const MeteringPointType = {
  ELECTRICITY: 'ELECTRICITY',
  HEAT: 'HEAT',
  COOLING: 'COOLING',
  PRODUCTION: 'PRODUCTION',
  WATER: 'WATER'
} as const

export type MeteringPointType = (typeof MeteringPointType)[keyof typeof MeteringPointType]
// ...
export const SubMeterCategory = {
  OPPVARMING: 'OPPVARMING',
  VENTILASJON: 'VENTILASJON',
  KJOLING: 'KJOLING',
  BELYSNING: 'BELYSNING',
  STIKKONTAKT: 'STIKKONTAKT',
  TAPPEVANN: 'TAPPEVANN',
  HEIS: 'HEIS',
  ELBIL_LADING: 'ELBIL_LADING',
  IT_DATAROM: 'IT_DATAROM',
  VARMEPUMPE: 'VARMEPUMPE',
  SOLCELLE: 'SOLCELLE',
  SNOSMELTING: 'SNOSMELTING',
  FORDELING: 'FORDELING',
  FELLES: 'FELLES',
  TEKNISK: 'TEKNISK',
  PROSESS: 'PROSESS',
  ANNET: 'ANNET'
} as const

export type SubMeterCategory = (typeof SubMeterCategory)[keyof typeof SubMeterCategory]


export const SubMeterSubCategory = {
  GULVVARME: 'GULVVARME',
  RADIATOR: 'RADIATOR',
  KONVEKTOR: 'KONVEKTOR',
  AEROTEMPER: 'AEROTEMPER',
  SNOSMELTING_VARMEKABLER: 'SNOSMELTING_VARMEKABLER',
  VARMEBATTERI: 'VARMEBATTERI',
  VIFTE: 'VIFTE',
  VARMEBATTERI_VENT: 'VARMEBATTERI_VENT',
  KJOLEBATTERI: 'KJOLEBATTERI',
  ROTOR_GJENVINNER: 'ROTOR_GJENVINNER',
  CHILLER: 'CHILLER',
  DX_AGGREGAT: 'DX_AGGREGAT',
  TORRKJOLER: 'TORRKJOLER',
  PROSESSKJOLING: 'PROSESSKJOLING',
  VP_PRODUKSJON: 'VP_PRODUKSJON',
  VP_KALD_SIDE: 'VP_KALD_SIDE',
  VP_VARM_SIDE: 'VP_VARM_SIDE',
  KJOKKEN: 'KJOKKEN',
  LAB: 'LAB',
  VERKSTED: 'VERKSTED',
  KONTOR: 'KONTOR',
  PROSESSUTSTYR: 'PROSESSUTSTYR',
  INNVENDIG_BELYSNING: 'INNVENDIG_BELYSNING',
  UTVENDIG_BELYSNING: 'UTVENDIG_BELYSNING',
  NODBELYSNING: 'NODBELYSNING',
  IT_LAST: 'IT_LAST',
  UPS: 'UPS',
  IT_KJOLING: 'IT_KJOLING',
  AUTOMASJON: 'AUTOMASJON',
  STYRING: 'STYRING',
  MIDLERTIDIG_STROM: 'MIDLERTIDIG_STROM',
  VVB: 'VVB',
  SIRKULASJONSPUMPE: 'SIRKULASJONSPUMPE',
  PERSONHEIS: 'PERSONHEIS',
  VAREHEIS: 'VAREHEIS',
  RULLETRAPP: 'RULLETRAPP',
  NORMALLADER: 'NORMALLADER',
  HURTIGLADER: 'HURTIGLADER',
  PV_PANEL: 'PV_PANEL',
  INVERTER: 'INVERTER',
  BATTERILAGRING: 'BATTERILAGRING'
} as const

export type SubMeterSubCategory = (typeof SubMeterSubCategory)[keyof typeof SubMeterSubCategory]